refactor: make Endpoint be the actual raw HTTP client
