+++ test/BaseClient.test.js
@@ -3,7 +3,7 @@ const { text, urlencoded } = require('body-parser')
 const fetch = require('nodeify-fetch')
 const { describe, it } = require('mocha')
 const Endpoint = require('../Endpoint')
-const RawQuery = require('../RawQuery')
+const BaseClient = require('../BaseClient')
 const withServer = require('./support/withServer')
 
 const simpleAskQuery = 'ASK {}'
@@ -11,547 +11,15 @@ const simpleConstructQuery = 'CONSTRUCT {?s ?p ?o} WHERE {?s ?p ?o}'
 const simpleSelectQuery = 'SELECT * WHERE {?s ?p ?o}'
 const simpleUpdateQuery = 'INSERT {<http://example.org/subject> <http://example.org/predicate> "object"} WHERE {}'
 
-describe('RawQuery', () => {
+describe('BaseClient', () => {
   it('should be a constructor', () => {
-    strictEqual(typeof RawQuery, 'function')
-  })
-
-  describe('.get', () => {
-    it('should be a method', () => {
-      const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
-
-      strictEqual(typeof query.get, 'function')
-    })
-
-    it('should async return a response object', async () => {
-      await withServer(async server => {
-        server.app.get('/', async (req, res) => {
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        const res = await query.get(simpleSelectQuery)
-
-        strictEqual(typeof res, 'object')
-        strictEqual(typeof res.text, 'function')
-      })
-    })
-
-    it('should send a GET request to the endpointUrl', async () => {
-      await withServer(async server => {
-        let called = false
-
-        server.app.get('/', async (req, res) => {
-          called = true
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleSelectQuery)
-
-        strictEqual(called, true)
-      })
-    })
-
-    it('should send the query string as query parameter', async () => {
-      await withServer(async server => {
-        let parameter = null
-
-        server.app.get('/', async (req, res) => {
-          parameter = req.query.query
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleSelectQuery)
-
-        strictEqual(parameter, simpleSelectQuery)
-      })
-    })
-
-    it('should keep existing query params', async () => {
-      await withServer(async server => {
-        let parameters = null
-        const key = 'auth_token'
-        const value = '12345'
-
-        server.app.get('/', async (req, res) => {
-          parameters = req.query
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl: `${endpointUrl}?${key}=${value}`, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleSelectQuery)
-
-        strictEqual(parameters[key], value)
-      })
-    })
-
-    it('should merge the headers given in the method call', async () => {
-      await withServer(async server => {
-        let header = null
-        const value = 'Bearer foo'
-
-        server.app.get('/', async (req, res) => {
-          header = req.headers.authorization
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleConstructQuery, {
-          headers: {
-            authorization: value
-          }
-        })
-
-        strictEqual(header, value)
-      })
-    })
-
-    it('should prioritize the headers from the method call', async () => {
-      await withServer(async server => {
-        let header = null
-        const value = 'Bearer foo'
-
-        server.app.get('/', async (req, res) => {
-          header = req.headers.authorization
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({
-          endpointUrl,
-          fetch,
-          headers: {
-            authorization: 'Bearer bar'
-          }
-        })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleConstructQuery, {
-          headers: {
-            authorization: value
-          }
-        })
-
-        strictEqual(header, value)
-      })
-    })
-
-    it('should use the updateUrl and update param if update is true', async () => {
-      await withServer(async server => {
-        let parameters = null
-
-        server.app.get('/', async (req, res) => {
-          parameters = req.query
-
-          res.end()
-        })
-
-        const updateUrl = await server.listen()
-
-        const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleSelectQuery, { update: true })
-
-        strictEqual(parameters.update, simpleSelectQuery)
-      })
-    })
-
-    it('should keep existing update query params', async () => {
-      await withServer(async server => {
-        let parameters = null
-        const key = 'auth_token'
-        const value = '12345'
-
-        server.app.get('/', async (req, res) => {
-          parameters = req.query
-
-          res.end()
-        })
-
-        const updateUrl = await server.listen()
-
-        const endpoint = new Endpoint({ updateUrl: `${updateUrl}?${key}=${value}`, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.get(simpleSelectQuery, { update: true })
-
-        strictEqual(parameters[key], value)
-      })
-    })
-  })
-
-  describe('.postDirect', () => {
-    it('should be a method', () => {
-      const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
-
-      strictEqual(typeof query.postDirect, 'function')
-    })
-
-    it('should async return a response object', async () => {
-      await withServer(async server => {
-        server.app.post('/', async (req, res) => {
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        const res = await query.postDirect(simpleSelectQuery)
-
-        strictEqual(typeof res, 'object')
-        strictEqual(typeof res.text, 'function')
-      })
-    })
-
-    it('should send a POST request to the endpointUrl', async () => {
-      await withServer(async server => {
-        let called = false
-
-        server.app.post('/', async (req, res) => {
-          called = true
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postDirect(simpleSelectQuery)
-
-        strictEqual(called, true)
-      })
-    })
-
-    it('should send a content type header with the value application/sparql-query & charset utf-8', async () => {
-      await withServer(async server => {
-        let contentType = null
-
-        server.app.post('/', async (req, res) => {
-          contentType = req.headers['content-type']
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postDirect(simpleSelectQuery)
-
-        strictEqual(contentType, 'application/sparql-query; charset=utf-8')
-      })
-    })
-
-    it('should send the query string in the request body', async () => {
-      await withServer(async server => {
-        let content = null
-
-        server.app.post('/', text({ type: '*/*' }), async (req, res) => {
-          content = req.body
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postDirect(simpleSelectQuery)
-
-        strictEqual(content, simpleSelectQuery)
-      })
-    })
-
-    it('should merge the headers given in the method call', async () => {
-      await withServer(async server => {
-        let header = null
-        const value = 'Bearer foo'
-
-        server.app.post('/', async (req, res) => {
-          header = req.headers.authorization
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postDirect(simpleConstructQuery, {
-          headers: {
-            authorization: value
-          }
-        })
-
-        strictEqual(header, value)
-      })
-    })
-
-    it('should prioritize the headers from the method call', async () => {
-      await withServer(async server => {
-        let header = null
-        const value = 'Bearer foo'
-
-        server.app.post('/', async (req, res) => {
-          header = req.headers.authorization
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({
-          endpointUrl,
-          fetch,
-          headers: {
-            authorization: 'Bearer bar'
-          }
-        })
-        const query = new RawQuery({ endpoint })
-
-        await query.postDirect(simpleConstructQuery, {
-          headers: {
-            authorization: value
-          }
-        })
-
-        strictEqual(header, value)
-      })
-    })
-
-    it('should use the updateUrl if update is true', async () => {
-      await withServer(async server => {
-        let called = false
-
-        server.app.post('/', async (req, res) => {
-          called = true
-
-          res.end()
-        })
-
-        const updateUrl = await server.listen()
-
-        const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postDirect(simpleSelectQuery, { update: true })
-
-        strictEqual(called, true)
-      })
-    })
-  })
-
-  describe('.postUrlencoded', () => {
-    it('should be a method', () => {
-      const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
-
-      strictEqual(typeof query.postUrlencoded, 'function')
-    })
-
-    it('should async return a response object', async () => {
-      await withServer(async server => {
-        server.app.post('/', async (req, res) => {
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        const res = await query.postUrlencoded(simpleSelectQuery)
-
-        strictEqual(typeof res, 'object')
-        strictEqual(typeof res.text, 'function')
-      })
-    })
-
-    it('should send a POST request to the endpointUrl', async () => {
-      await withServer(async server => {
-        let called = false
-
-        server.app.post('/', async (req, res) => {
-          called = true
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postUrlencoded(simpleSelectQuery)
-
-        strictEqual(called, true)
-      })
-    })
-
-    it('should send a content type header with the value application/x-www-form-urlencoded', async () => {
-      await withServer(async server => {
-        let contentType = null
-
-        server.app.post('/', async (req, res) => {
-          contentType = req.headers['content-type']
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postUrlencoded(simpleSelectQuery)
-
-        strictEqual(contentType, 'application/x-www-form-urlencoded')
-      })
-    })
-
-    it('should send the query string urlencoded in the request body', async () => {
-      await withServer(async server => {
-        let parameter = null
-
-        server.app.post('/', urlencoded({ extended: true }), async (req, res) => {
-          parameter = req.body.query
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postUrlencoded(simpleSelectQuery)
-
-        strictEqual(parameter, simpleSelectQuery)
-      })
-    })
-
-    it('should merge the headers given in the method call', async () => {
-      await withServer(async server => {
-        let header = null
-        const value = 'Bearer foo'
-
-        server.app.post('/', async (req, res) => {
-          header = req.headers.authorization
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postUrlencoded(simpleConstructQuery, {
-          headers: {
-            authorization: value
-          }
-        })
-
-        strictEqual(header, value)
-      })
-    })
-
-    it('should prioritize the headers from the method call', async () => {
-      await withServer(async server => {
-        let header = null
-        const value = 'Bearer foo'
-
-        server.app.post('/', async (req, res) => {
-          header = req.headers.authorization
-
-          res.end()
-        })
-
-        const endpointUrl = await server.listen()
-
-        const endpoint = new Endpoint({
-          endpointUrl,
-          fetch,
-          headers: {
-            authorization: 'Bearer bar'
-          }
-        })
-        const query = new RawQuery({ endpoint })
-
-        await query.postUrlencoded(simpleConstructQuery, {
-          headers: {
-            authorization: value
-          }
-        })
-
-        strictEqual(header, value)
-      })
-    })
-
-    it('should use the updateUrl if update is true', async () => {
-      await withServer(async server => {
-        let called = false
-
-        server.app.post('/', async (req, res) => {
-          called = true
-
-          res.end()
-        })
-
-        const updateUrl = await server.listen()
-
-        const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
-
-        await query.postUrlencoded(simpleSelectQuery, { update: true })
-
-        strictEqual(called, true)
-      })
-    })
+    strictEqual(typeof BaseClient, 'function')
   })
 
   describe('.ask', () => {
     it('should be a method', () => {
       const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
+      const query = new BaseClient({ endpoint })
 
       strictEqual(typeof query.ask, 'function')
     })
@@ -565,7 +33,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         const res = await query.ask(simpleAskQuery)
 
@@ -587,7 +55,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleAskQuery)
 
@@ -608,7 +76,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleSelectQuery)
 
@@ -631,7 +99,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl: `${endpointUrl}?${key}=${value}`, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleAskQuery)
 
@@ -652,7 +120,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleAskQuery)
 
@@ -674,7 +142,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleAskQuery, {
           headers: {
@@ -706,7 +174,7 @@ describe('RawQuery', () => {
             authorization: 'Bearer bar'
           }
         })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleAskQuery, {
           headers: {
@@ -731,7 +199,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.ask(simpleAskQuery, { operation: 'postUrlencoded' })
 
@@ -743,7 +211,7 @@ describe('RawQuery', () => {
   describe('.construct', () => {
     it('should be a method', () => {
       const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
+      const query = new BaseClient({ endpoint })
 
       strictEqual(typeof query.construct, 'function')
     })
@@ -757,7 +225,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         const res = await query.construct(simpleConstructQuery)
 
@@ -779,7 +247,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery)
 
@@ -800,7 +268,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery)
 
@@ -823,7 +291,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl: `${endpointUrl}?${key}=${value}`, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery)
 
@@ -844,7 +312,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery)
 
@@ -866,7 +334,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery, {
           headers: {
@@ -898,7 +366,7 @@ describe('RawQuery', () => {
             authorization: 'Bearer bar'
           }
         })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery, {
           headers: {
@@ -923,7 +391,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.construct(simpleConstructQuery, { operation: 'postUrlencoded' })
 
@@ -935,7 +403,7 @@ describe('RawQuery', () => {
   describe('.select', () => {
     it('should be a method', () => {
       const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
+      const query = new BaseClient({ endpoint })
 
       strictEqual(typeof query.select, 'function')
     })
@@ -949,7 +417,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         const res = await query.select(simpleSelectQuery)
 
@@ -971,7 +439,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery)
 
@@ -992,7 +460,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery)
 
@@ -1015,7 +483,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl: `${endpointUrl}?${key}=${value}`, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery)
 
@@ -1036,7 +504,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery)
 
@@ -1058,7 +526,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery, {
           headers: {
@@ -1090,7 +558,7 @@ describe('RawQuery', () => {
             authorization: 'Bearer bar'
           }
         })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery, {
           headers: {
@@ -1115,7 +583,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleSelectQuery, { operation: 'postUrlencoded' })
 
@@ -1127,7 +595,7 @@ describe('RawQuery', () => {
   describe('.update', () => {
     it('should be a method', () => {
       const endpoint = new Endpoint({ fetch })
-      const query = new RawQuery({ endpoint })
+      const query = new BaseClient({ endpoint })
 
       strictEqual(typeof query.update, 'function')
     })
@@ -1141,7 +609,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         const res = await query.update(simpleUpdateQuery)
 
@@ -1163,7 +631,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery)
 
@@ -1186,7 +654,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl: `${updateUrl}?${key}=${value}`, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery)
 
@@ -1207,7 +675,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery)
 
@@ -1228,7 +696,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery)
 
@@ -1249,7 +717,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery)
 
@@ -1271,7 +739,7 @@ describe('RawQuery', () => {
         const updateUrl = await server.listen()
 
         const endpoint = new Endpoint({ updateUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery, {
           headers: {
@@ -1303,7 +771,7 @@ describe('RawQuery', () => {
             authorization: 'Bearer bar'
           }
         })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.update(simpleUpdateQuery, {
           headers: {
@@ -1328,7 +796,7 @@ describe('RawQuery', () => {
         const endpointUrl = await server.listen()
 
         const endpoint = new Endpoint({ endpointUrl, fetch })
-        const query = new RawQuery({ endpoint })
+        const query = new BaseClient({ endpoint })
 
         await query.select(simpleUpdateQuery, { operation: 'postDirect' })
